Finalizado os EndPoint de atualizar e deletar produtos.

+++ src/main/java/onecenter/com/br/ecommerce/repository/mapper/ProdutosRowMapper.java
@@ -1,5 +1,6 @@
 package onecenter.com.br.ecommerce.repository.mapper;
 
+import onecenter.com.br.ecommerce.entity.produtos.categoria.CategoriaEntity;
 import onecenter.com.br.ecommerce.entity.produtos.ProdutosEntity;
 import org.springframework.jdbc.core.RowMapper;
 
@@ -11,7 +12,7 @@ public class ProdutosRowMapper implements RowMapper<ProdutosEntity> {
         ProdutosEntity produtos = new ProdutosEntity();
         produtos.setId_produto(rs.getInt(1));
         produtos.setNome(rs.getString(2));
-        produtos.setPreco(rs.getFloat(3));
+        produtos.setPreco(rs.getDouble(3));
         produtos.setProduto_imagem(rs.getString(4));
         produtos.setId_categoria(rs.getInt(5));
         return produtos;
